Enum cranelift_isle::ast::Pattern

pub enum Pattern {
    Var {
        var: Ident,
        pos: Pos,
    },
    BindPattern {
        var: Ident,
        subpat: Box<Pattern>,
        pos: Pos,
    },
    ConstInt {
        val: i128,
        pos: Pos,
    },
    ConstPrim {
        val: Ident,
        pos: Pos,
    },
    Term {
        sym: Ident,
        args: Vec<Pattern>,
        pos: Pos,
    },
    Wildcard {
        pos: Pos,
    },
    And {
        subpats: Vec<Pattern>,
        pos: Pos,
    },
    MacroArg {
        index: usize,
        pos: Pos,
    },
}

A pattern: the left-hand side of a rule.

Variants

Var

Fields

var: Ident

pos: Pos

A mention of a variable.

Equivalent either to a binding (which can be emulated with BindPattern with a Pattern::Wildcard subpattern), if this is the first mention of the variable, in order to capture its value; or else a match of the already-captured value. This disambiguation happens when we lower ast nodes to sema nodes as we resolve bound variable names.

BindPattern

Fields

var: Ident

subpat: Box<Pattern>

pos: Pos

An operator that binds a variable to a subterm and matches the subpattern.

ConstInt

Fields

val: i128

pos: Pos

An operator that matches a constant integer value.

ConstPrim

Fields

val: Ident

pos: Pos

An operator that matches an external constant value.

Term

Fields

sym: Ident

args: Vec<Pattern>

pos: Pos

An application of a type variant or term.

Wildcard

Fields

pos: Pos

An operator that matches anything.

And

Fields

subpats: Vec<Pattern>

pos: Pos

N sub-patterns that must all match.

MacroArg

Fields

index: usize

pos: Pos

Internal use only: macro argument in a template.

Implementations

impl Pattern

pub fn root_term(&self) -> Option<&Ident>

pub fn terms(&self, f: &mut dyn FnMut(Pos, &Ident))

Call f for each of the terms in this pattern.

pub fn make_macro_template(&self, macro_args: &[Ident]) -> Pattern

pub fn subst_macro_args(&self, macro_args: &[Pattern]) -> Option<Pattern>

pub fn pos(&self) -> Pos