Enum cranelift_isle::ir::PatternInst

pub enum PatternInst {
    MatchEqual {
        a: Value,
        b: Value,
        ty: TypeId,
    },
    MatchInt {
        input: Value,
        ty: TypeId,
        int_val: i128,
    },
    MatchPrim {
        input: Value,
        ty: TypeId,
        val: Sym,
    },
    MatchVariant {
        input: Value,
        input_ty: TypeId,
        arg_tys: Vec<TypeId>,
        variant: VariantId,
    },
    Expr {
        seq: ExprSequence,
        output: Value,
        output_ty: TypeId,
    },
    Extract {
        infallible: bool,
        inputs: Vec<Value>,
        input_tys: Vec<TypeId>,
        output_tys: Vec<TypeId>,
        term: TermId,
        multi: bool,
    },
    Arg {
        index: usize,
        ty: TypeId,
    },
}

A single Pattern instruction.

Variants

MatchEqual

Fields

a: Value

The first value.

b: Value

The second value.

ty: TypeId

The type of the values.

Match a value as equal to another value. Produces no values.

MatchInt

Fields

input: Value

The value to match on.

ty: TypeId

The value’s type.

int_val: i128

The integer to match against the value.

Try matching the given value as the given integer. Produces no values.

MatchPrim

Fields

input: Value

The value to match on.

ty: TypeId

The type of the value.

val: Sym

The primitive to match against the value.

Try matching the given value as the given constant. Produces no values.

MatchVariant

Fields

input: Value

The value to match on.

input_ty: TypeId

The type of the value.

arg_tys: Vec<TypeId>

The types of values produced upon a successful match.

variant: VariantId

The value type’s variant that we are matching against.

Try matching the given value as the given variant, producing |arg_tys| values as output.

Expr

Fields

seq: ExprSequence

The expression to evaluate.

output: Value

The value produced by the expression.

output_ty: TypeId

The type of the output value.

Evaluate an expression and provide the given value as the result of this match instruction. The expression has access to the pattern-values up to this point in the sequence.

Extract

Fields

infallible: bool

Whether this extraction is infallible or not. false comes before true, so fallible nodes come first.

inputs: Vec<Value>

The value to extract, followed by polarity extractor args.

input_tys: Vec<TypeId>

The types of the inputs.

output_tys: Vec<TypeId>

The types of the output values produced upon a successful match.

term: TermId

This extractor’s term.

multi: bool

Is this a multi-extractor?

Invoke an extractor, taking the given values as input (the first is the value to extract, the other are the Input-polarity extractor args) and producing an output value for each Output-polarity extractor arg.

Arg

Fields

index: usize

The index of the argument to get.

ty: TypeId

The type of the argument.

Get the Nth input argument, which corresponds to the Nth field of the root term.