Crate cranelift_isle

ISLE: Instruction Selection / Lowering Expressions

ISLE is a domain specific language (DSL) for instruction selection and lowering clif instructions to vcode’s MachInsts in Cranelift.

ISLE is a statically-typed term-rewriting language. You define rewriting rules that map input terms (clif instructions) into output terms (MachInsts). These rules get compiled down into Rust source test that uses a tree of match expressions that is as good or better than what you would have written by hand.

Modules

ast

Abstract syntax tree (AST) created from parsed ISLE.

codegen

Generate Rust code from a series of Sequences.

compile

Compilation process, from AST to Sema to Sequences of Insts.

error

Error types.

ir

Lowered matching IR.

lexer

Lexer for the ISLE language.

overlap

Overlap detection for rules in ISLE.

parser

Parser for ISLE language.

sema

Semantic analysis.

trie

Trie construction.

Macros

log

Log a compiler-internal message for debugging purposes.

Structs

StableMap

A wrapper around a HashMap which prevents accidentally observing the non-deterministic iteration order.